ENH: Update default tools location to toolsOctober2014
This allows to use the --shiftNeg option in dtiestim that shifts negative eigenvalues to only have strictly positive eigenvalues.
ITKTransformTools had a bug when running d2f because of the new implementation of the ITKTransformFileReader(Template) in ITK4.5 and above. The new version of ITKTransformTools in toolsOctober2014 also correct this issue.