ENH: Processing tools directory is suppose to be in the current project's tools directory
The processing tools directory in which the executables are found used to be somewhere on the system. The script user could change the directory but it was pointing by default to the most recent tools* directory in /rodent/bin_linux64. Now, to make it easier to archive projects, we expect the user to copy the tools in the project directory, under path/to/project/tools/tools${Month}${year}. The configuration script will attempt to find that folder and configure the script using it. The user can still manually modify what the configuration script finds.